Dialogue System/Cutscene System implementation + AoE Attack + Misc
The dialogue box is attatched to the Event System and runs at the beginning every single time. For now, it completely disables the player and boss, to prevent any background scripts from doing random things. If it gets to tedious to go through it each time, I can disable it in the future.