First pass at dodge, cleaned up a tiny bit of my code, added some assets, and refreshed project on my end
My Unity continously refused to open, so I made a new project and overwrote most of it with the files from this project. Due to the gitignore, not too many files that were changed during this process should carry over to github, but I checked around and it seems like most everything is fine.